requirements.txt
pillow
mss
pytesseract
python-dotenv
requests
keyboard
anthropic
google-generativeai
openai>=1.30.0
customtkinter
